♻️ refactor: view 폴더 위치 변경 및 메소드, 변수명 수정
