+++ docs/README.md
@@ -1,7 +1,7 @@
 # 🚗 자동차 경주 게임 🚗
 
 ## ➊ 사용자에게 경주할 자동차 이름을 입력받는다.
-- [x] 사용자로부터 자동차 이름을 입력받아 반환 → `getCarName()`
+- [x] 사용자로부터 자동차 이름을 입력받아 반환 → `getCarNames()`
 - [ ] 자동차 이름이 입력 형식에 맞는지 확인 → `validateCarName()`
     - [ ] 고려 사항 👀
         - [x] 자동차 이름은 1~5글자 `예외) okxooxoo,poby`
@@ -25,7 +25,7 @@
 - [x] 0~9 사이의 랜덤값을 생성하여 4 이상이면 `true` 반환 → `isMove()`
 - [x] `ismove()`의 결과에 따라 자동차 전진 OR 정지 → `moveCars()`
 - [x] 한 차수가 끝나면 모든 자동차의 이동 거리 출력 → `printCars()`
-    - [x] 자동차의 이동 거리에 따른 `-` 문자열 반환 → `getLine()`
+    - [x] 자동차의 이동 거리에 따른 `-` 문자열 반환 → `getLines()`
 
 ## ➍ 최종 우승자를 출력한다.
 - [x] 최종 우승자를 판단하고 우승자 이름 반환 → `getWinners()`
