+++ src/view/InputView.js
@@ -1,10 +1,10 @@
 import { Console } from '@woowacourse/mission-utils';
-import { splitByComma } from '../utils/splitByComma.js';
-import { validateCarName, validateTryNumber } from '../utils/inputValidator.js';
+import { splitByComma } from '../../utils/splitByComma.js';
+import { validateCarName, validateTryNumber } from '../../utils/inputValidator.js';
 
 class InputView {
   
-  static async getCarName() {
+  static async getCarNames() {
     const carInput = await Console.readLineAsync("경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)\n");
     validateCarName(carInput);
     const carNames = splitByComma(carInput);
